feat: Implement Binder, communicate with service when session is discovered to start playback
